# target/generated-sources/swagger/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "Documentación de los servicios Rest sobre HTTP expuestos por JBoss FUSE para manejo de pedidos, productos y recepciones del proyecto Customer Experience.",
    "version" : "1.0.0",
    "title" : "Newsan - Customer Experience",
    "contact" : {
      "name" : "Ingenia CA",
      "email" : "mail@ingenia.la"
    }
  },
  "basePath" : "/custexp",
...
  "schemes" : [ "https", "http" ],
  "paths" : {
    "/order/{type}/{id}" : {
      "get" : {
        "tags" : [ "orders" ],
        "summary" : "find an order in EBS",
        "description" : "Consulta un pedido en EBS",
        "operationId" : "findOrder",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "pass an order id for looking up specific order",
          "required" : true,
          "type" : "number"
        }, {
          "name" : "type",
          "in" : "path",
          "description" : "pass an order type for looking up specific order",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "unit",
          "in" : "query",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Successful Operation",
            "schema" : {
              "type" : "object",
              "properties" : {
                "data" : {
                  "type" : "object",
                  "properties" : { }
                }
              }
            }
          },
          "404" : {
            "description" : "Order not Found",
            "schema" : {
              "type" : "object",
              "properties" : {
                "errors" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/definitions/Error"
                  }
                }
              }
            }
          },
          "500" : {
            "description" : "Internal Server Error",
            "schema" : {
              "type" : "object",
              "properties" : {
                "errors" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/definitions/Error"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/order" : {
      "post" : {
        "tags" : [ "orders" ],
        "summary" : "create an order in EBS",
        "description" : "Crea un pedido en EBS",
        "operationId" : "createOrder",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "order",
          "description" : "The order to create",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/ApiOrder"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Successful Operation",
            "schema" : {
              "type" : "object",
              "properties" : {
                "data" : {
                  "type" : "object",
                  "properties" : { }
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid Parameters",
            "schema" : {
              "type" : "object",
              "properties" : {
                "errors" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/definitions/Error"
                  }
                }
              }
            }
          },
          "500" : {
            "description" : "Internal Server Error",
            "schema" : {
              "type" : "object",
              "properties" : {
                "errors" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/definitions/Error"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/product/{sku}/price" : {
      "get" : {
        "tags" : [ "products" ],
        "summary" : "find a product price in EBS",
        "description" : "Consulta el precio de un producto en EBS",
        "operationId" : "findProductPrice",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "headerName",
          "in" : "query",
          "description" : "pass a product header name for looking up specific product",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "sku",
          "in" : "path",
          "description" : "pass a product sku for looking up specific product",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Successful Operation",
            "schema" : {
              "type" : "object",
              "properties" : {
                "data" : {
                  "type" : "object",
                  "properties" : { }
                }
              }
            }
          },
          "404" : {
            "description" : "Product not Found",
            "schema" : {
              "type" : "object",
              "properties" : {
                "errors" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/definitions/Error"
                  }
                }
              }
            }
          },
          "500" : {
            "description" : "Internal Server Error",
            "schema" : {
              "type" : "object",
              "properties" : {
                "errors" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/definitions/Error"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/product/{sku}/stock" : {
      "get" : {
        "tags" : [ "products" ],
        "summary" : "find a product stock in EBS",
        "description" : "Consulta el stock de un producto en EBS",
        "operationId" : "findProductStock",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "organization",
          "in" : "query",
          "description" : "pass a product organization code for filtering SKU results",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "subinventory",
          "in" : "query",
          "description" : "pass a product subinventory code for filtering SKU results",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "sku",
          "in" : "path",
          "description" : "pass a product sku for looking up specific product",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Successful Operation",
            "schema" : {
              "type" : "object",
              "properties" : {
                "data" : {
                  "type" : "object",
                  "properties" : { }
                }
              }
            }
          },
          "404" : {
            "description" : "Product not Found",
            "schema" : {
              "type" : "object",
              "properties" : {
                "errors" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/definitions/Error"
                  }
                }
              }
            }
          },
          "500" : {
            "description" : "Internal Server Error",
            "schema" : {
              "type" : "object",
              "properties" : {
                "errors" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/definitions/Error"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/product/{organization}" : {
      "get" : {
        "tags" : [ "products" ],
        "summary" : "list EBS products by organization filtered by last update date",
        "description" : "Consulta productos en EBS en base a la organización y ultima fecha de actualización.",
        "operationId" : "listProducts",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "organization",
          "in" : "path",
          "description" : "pass a product organization",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "lastUpdated",
          "in" : "query",
          "description" : "pass last updated date to filter products",
          "required" : true,
          "type" : "string",
          "pattern" : "\\d{4}-\\d{2}-\\d{2}"
        } ],
        "responses" : {
          "200" : {
            "description" : "Successful Operation",
            "schema" : {
              "type" : "object",
              "properties" : {
                "data" : {
                  "type" : "object",
                  "properties" : { }
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid Parameters",
            "schema" : {
              "type" : "object",
              "properties" : {
                "errors" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/definitions/Error"
                  }
                }
              }
            }
          },
          "404" : {
            "description" : "Product not Found",
            "schema" : {
              "type" : "object",
              "properties" : {
                "errors" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/definitions/Error"
                  }
                }
              }
            }
          },
          "500" : {
            "description" : "Internal Server Error",
            "schema" : {
              "type" : "object",
              "properties" : {
                "errors" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/definitions/Error"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/product/{organization}/components" : {
      "get" : {
        "tags" : [ "products" ],
        "summary" : "list EBS product components by organization filtered by last update date",
        "description" : "Consulta despiece de productos en EBS en base a la organización y ultima fecha de actualización.",
        "operationId" : "listProductComponents",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "organization",
          "in" : "path",
          "description" : "pass a product organization",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "lastUpdated",
          "in" : "query",
          "description" : "pass last updated date to filter products",
          "required" : true,
          "type" : "string",
          "pattern" : "\\d{4}-\\d{2}-\\d{2}"
        } ],
        "responses" : {
          "200" : {
            "description" : "Successful Operation",
            "schema" : {
              "type" : "object",
              "properties" : {
                "data" : {
                  "type" : "object",
                  "properties" : { }
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid Parameters",
            "schema" : {
              "type" : "object",
              "properties" : {
                "errors" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/definitions/Error"
                  }
                }
              }
            }
          },
          "404" : {
            "description" : "Product not Found",
            "schema" : {
              "type" : "object",
              "properties" : {
                "errors" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/definitions/Error"
                  }
                }
              }
            }
          },
          "500" : {
            "description" : "Internal Server Error",
            "schema" : {
              "type" : "object",
              "properties" : {
                "errors" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/definitions/Error"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/customer" : {
      "get" : {
        "tags" : [ "customers" ],
        "summary" : "list EBS customers filtered by last update date",
        "description" : "Consulta clientes en EBS en base a la ultima fecha de actualización.",
        "operationId" : "listCustomers",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "lastUpdated",
          "in" : "query",
          "description" : "pass last updated date to filter products",
          "required" : true,
          "type" : "string",
          "pattern" : "\\d{4}-\\d{2}-\\d{2}"
        } ],
        "responses" : {
          "200" : {
            "description" : "Successful Operation",
            "schema" : {
              "type" : "object",
              "properties" : {
                "data" : {
                  "type" : "object",
                  "properties" : { }
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid Parameters",
            "schema" : {
              "type" : "object",
              "properties" : {
                "errors" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/definitions/Error"
                  }
                }
              }
            }
          },
          "404" : {
            "description" : "Customer not Found",
            "schema" : {
              "type" : "object",
              "properties" : {
                "errors" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/definitions/Error"
                  }
                }
              }
            }
          },
          "500" : {
            "description" : "Internal Server Error",
            "schema" : {
              "type" : "object",
              "properties" : {
                "errors" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/definitions/Error"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "definitions" : {
...
    "Error" : {
      "type" : "object",
      "properties" : {
        "code" : {
          "type" : "string"
        },
        "message" : {
          "type" : "string"
        }
      }
    },
    "ApiOrder" : {
      "type" : "object",
      "required" : [ "clientAccount", "clientBillRef", "clientShipRef", "lines", "orderMIB", "organizationCode", "paymentTerm", "priceListName", "salesRep", "shipFrom", "source", "typeName" ],
      "properties" : {
        "typeName" : {
          "type" : "string",
          "description" : "Tipo de pedido"
        },
        "clientAccount" : {
          "type" : "string",
          "description" : "Id de cliente"
        },
        "clientShipRef" : {
          "type" : "string",
          "description" : "Identificador de referencia de envio del cliente"
        },
        "clientBillRef" : {
          "type" : "string",
          "description" : "Identificador de referencia de facturacion del cliente"
        },
        "source" : {
          "type" : "integer",
          "description" : "Id de origen del pedido",
          "minimum" : 1
        },
        "bookedFlag" : {
          "type" : "string",
          "description" : "Código de estado de inicio del pedido",
          "default" : "N"
        },
        "priceListName" : {
          "type" : "string",
          "description" : "Nombre de la lista de precios"
        },
        "flowStatusCode" : {
          "type" : "string",
          "description" : "Código del estado del pedido",
          "default" : "BOOKED"
        },
        "organizationCode" : {
          "type" : "string",
          "description" : "Unidad operativa"
        },
        "salesRep" : {
          "type" : "string",
          "description" : "Corresponde al número del vendedor"
        },
        "currencyCode" : {
          "type" : "string",
          "description" : "Corresponde al código de moneda utilizada",
          "minLength" : 3,
          "maxLength" : 3,
          "default" : "ARS"
        },
        "orderMIB" : {
          "type" : "string",
          "description" : "Id del pedido generado en Salesforce"
        },
        "shipFrom" : {
          "type" : "string",
          "description" : "Origen del envío del pedido"
        },
        "ocExpirationDate" : {
          "type" : "string",
          "description" : "Fecha de vencimiento de la OC",
          "pattern" : "\\d{4}-\\d{2}-\\d{2}"
        },
        "paymentTerm" : {
          "type" : "integer",
          "description" : "Corresponde al término de pago del cliente",
          "minimum" : 1
        },
        "shippingInstructions" : {
          "type" : "string",
          "description" : "Corresponde a las instrucciones de envío"
        },
        "lines" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/ApiOrderLine"
          }
        }
      }
    },
    "ApiOrderLine" : {
      "type" : "object",
      "required" : [ "calculatePriceFlag", "lineMIB", "ordered", "paymentTerm", "shippingMethodCode", "sku", "type", "unitListPrice", "unitPrice" ],
      "properties" : {
        "sku" : {
          "type" : "string",
          "description" : "Id de SKU"
        },
        "type" : {
          "type" : "string",
          "description" : "Descripción de tipo de línea"
        },
        "unitPrice" : {
          "type" : "number",
          "format" : "float",
          "description" : "Precio unitario del producto",
          "minimum" : 1
        },
        "scheduleShipDate" : {
          "type" : "string",
          "description" : "Fecha programada de envío",
          "pattern" : "\\d{4}-\\d{2}-\\d{2}"
        },
        "ordered" : {
          "type" : "integer",
          "description" : "Cantidad solicitada por el cliente",
          "minimum" : 1
        },
        "calculatePriceFlag" : {
          "type" : "string",
          "description" : "Calcula el precio en función a la lista de precios seleccionada para la cabecera del pedido.",
          "maxLength" : 1,
          "default" : "Y"
        },
        "customerPO" : {
          "type" : "string",
          "description" : "Corresponde al número de OC del Cliente (el STA)"
        },
        "packingInstructions" : {
          "type" : "string",
          "description" : "Corresponde a las instrucciones de empaquetado del pedido"
        },
        "paymentTerm" : {
          "type" : "integer",
          "description" : "Corresponde al término de pago del cliente",
          "minimum" : 1
        },
        "lineMIB" : {
          "type" : "integer",
          "description" : "Id de la linea del pedido generado por Salesforce",
          "minimum" : 1
        },
        "shippingMethodCode" : {
          "type" : "string",
          "description" : "Código de método de envio"
        },
        "newPrice" : {
          "type" : "number",
          "format" : "float",
          "description" : "Nuevo precio con descuentos aplicados",
          "minimum" : 1
        },
        "unitListPrice" : {
          "type" : "number",
          "format" : "float",
          "description" : "Precio del artículo de la lista de precios",
          "minimum" : 1
        },
        "reasonCode" : {
          "type" : "string"
        },
        "context" : {
          "type" : "string"
        },
        "relatedTo" : {
          "type" : "string",
          "description" : "Utilizado para relacionar líneas sujetas al cumplimiento de otras"
        },
        "priceListName" : {
          "type" : "string"
        }
      }
    },
...
  }
}